# backend/rpg-backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: RPG DSL Backend API
  version: 1.0.0
  description: |
    OpenAPI-first backend for a scene-driven RPG DSL editor and runtime.
    Endpoints: validate/start/turn.

servers:
  - url: http://localhost:8080

tags:
  - name: DSL
  - name: Game

paths:
  /api/validate:
    post:
      tags: [DSL]
      summary: Validate DSL source (ANTLR syntax + semantic validation)
      operationId: validateDsl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateRequest"
      responses:
        "200":
          description: Validation result (diagnostics list; empty means OK)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateResponse"
        "400":
          description: Invalid request payload

  /api/start:
    post:
      tags: [Game]
      summary: Start a game from DSL source
      operationId: startGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRequest"
      responses:
        "200":
          description: Game started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartResponse"
        "422":
          description: DSL has validation errors (cannot start)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartResponse"
        "400":
          description: Invalid request payload

  /api/turn:
    post:
      tags: [Game]
      summary: Submit a player message (chat turn) using the full conversation history
      operationId: submitTurn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TurnRequest"
      responses:
        "200":
          description: Turn resolved from full chat history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TurnResponse"
        "400":
          description: Invalid request payload

components:
  schemas:
    ValidateRequest:
      type: object
      required: [source]
      properties:
        source:
          type: string
          minLength: 1
          description: Full DSL source text

    StartRequest:
      type: object
      required: [source]
      properties:
        source:
          type: string
          minLength: 1
          description: Full DSL source text to compile and start

    TurnRequest:
      type: object
      required: [source, history]
      properties:
        source:
          type: string
          minLength: 1
          description: Full DSL source text
        history:
          type: array
          description: Full conversation history in chronological order
          items:
            $ref: "#/components/schemas/ChatMessage"

    ValidateResponse:
      type: object
      required: [diagnostics]
      properties:
        diagnostics:
          type: array
          items:
            $ref: "#/components/schemas/Diagnostic"

    StartResponse:
      type: object
      required: [diagnostics]
      properties:
        text:
          type: string
          nullable: true
          description: Initial narration returned to client
        diagnostics:
          type: array
          items:
            $ref: "#/components/schemas/Diagnostic"
          description: Validation errors/warnings if start failed (or empty if OK)

    TurnResponse:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: Narration / response to show in chat

    ChatMessage:
      type: object
      required: [from, text]
      properties:
        from:
          type: string
          enum: [user, system]
        text:
          type: string

    Diagnostic:
      type: object
      required: [line, character, severity, message]
      properties:
        line:
          type: integer
          minimum: 0
          description: 1-based line number if available; 0 if unknown
        character:
          type: integer
          minimum: 0
          description: 1-based column number if available; 0 if unknown
        severity:
          type: string
          enum: [ERROR, WARN, INFO]
        code:
          type: string
          nullable: true
          description: Optional machine-readable code (e.g., SYNTAX_ERROR)
        message:
          type: string